// client/src/app/pages/human-resource/contract-end-employees/extend-contract/extendcontract.model.ts
export class ExtendContractDto {
    employeeId: string;
    createdById: string;
    startDate: Date;
    endDate: Date;
}

export class ContractExtentionLetterDto {
    todaysDate: Date;
    employeeName: string;
    position: string;
    previousStartDate: Date;
    previousEndDate: Date;
    pstartDate: Date;
    endDate: Date;
}